requirements.txt
pyyaml
flask
gevent
requests
pygithub
git+https://github.com/TwitchPlaysPokemon/pokecat.git
